+++ app.js
@@ -6,7 +6,6 @@ var logger = require('morgan');
 var mustacheExpress = require('mustache-express');
 
 var indexRouter = require('./routes/index');
-var oidcRouter = require('./routes/oidc');
 
 var app = express();
 
@@ -22,7 +21,6 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/', oidcRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
